Move `OutgoingMessage` type to `linera-base`
Allow the type to be accessed from `linera-execution` as well as from
user applications.
